+++ resources/js/Components/Jetstream/ActionMessage.vue
