+++ resources/js/Components/Jetstream/ApplicationLogo.vue
