+++ resources/js/Components/Jetstream/AuthenticationCard.vue
