+++ resources/js/Components/Jetstream/AuthenticationCardLogo.vue
