+++ resources/js/Components/Jetstream/Banner.vue
