+++ resources/js/Components/Jetstream/Checkbox.vue
