+++ resources/js/Components/Jetstream/DangerButton.vue
