+++ resources/js/Components/Jetstream/SectionTitle.vue
