+++ resources/js/Pages/Profile/Partials/TwoFactorAuthenticationForm.vue
@@ -2,14 +2,14 @@
 import { ref, computed, watch } from 'vue';
 import { Inertia } from '@inertiajs/inertia';
 import { useForm, usePage } from '@inertiajs/inertia-vue3';
-import JetActionSection from '@/Jetstream/ActionSection.vue';
-import JetButton from '@/Jetstream/Button.vue';
-import JetConfirmsPassword from '@/Jetstream/ConfirmsPassword.vue';
-import JetDangerButton from '@/Jetstream/DangerButton.vue';
-import JetInput from '@/Jetstream/Input.vue';
-import JetInputError from '@/Jetstream/InputError.vue';
-import JetLabel from '@/Jetstream/Label.vue';
-import JetSecondaryButton from '@/Jetstream/SecondaryButton.vue';
+import JetActionSection from '@/Components/Jetstream/ActionSection.vue';
+import JetButton from '@/Components/Jetstream/Button.vue';
+import JetConfirmsPassword from '@/Components/Jetstream/ConfirmsPassword.vue';
+import JetDangerButton from '@/Components/Jetstream/DangerButton.vue';
+import JetInput from '@/Components/Jetstream/Input.vue';
+import JetInputError from '@/Components/Jetstream/InputError.vue';
+import JetLabel from '@/Components/Jetstream/Label.vue';
+import JetSecondaryButton from '@/Components/Jetstream/SecondaryButton.vue';
 
 const props = defineProps({
   requiresConfirmation: Boolean,
