+++ resources/js/Pages/Profile/Partials/UpdatePasswordForm.vue
@@ -1,12 +1,12 @@
 <script setup>
 import { ref } from 'vue';
 import { useForm } from '@inertiajs/inertia-vue3';
-import JetActionMessage from '@/Jetstream/ActionMessage.vue';
-import JetButton from '@/Jetstream/Button.vue';
-import JetFormSection from '@/Jetstream/FormSection.vue';
-import JetInput from '@/Jetstream/Input.vue';
-import JetInputError from '@/Jetstream/InputError.vue';
-import JetLabel from '@/Jetstream/Label.vue';
+import JetActionMessage from '@/Components/Jetstream/ActionMessage.vue';
+import JetButton from '@/Components/Jetstream/Button.vue';
+import JetFormSection from '@/Components/Jetstream/FormSection.vue';
+import JetInput from '@/Components/Jetstream/Input.vue';
+import JetInputError from '@/Components/Jetstream/InputError.vue';
+import JetLabel from '@/Components/Jetstream/Label.vue';
 
 const passwordInput = ref(null);
 const currentPasswordInput = ref(null);
